use int64 stride everywhere

+++ mlx/backend/metal/kernels/indexing.h
@@ -8,7 +8,7 @@ template <typename IdxT, int NIDX>
 struct Indices {
   const array<const device IdxT*, NIDX> buffers;
   const constant int* shapes;
-  const constant size_t* strides;
+  const constant int64_t* strides;
   const constant bool* row_contiguous;
   const int ndim;
 };
